Accept params as dictionary (default) or list (optional)
This is useful for MCMC samplers and similar tools that pass the
parameters als list instead of dictionary.
Note that log_unnormalized_likelihood is the only function that allows
for list of parameters. The list gets converted there and calls the
other functions. This avoids needing to convert this everywhere.